requirements.txt
Flask==3.1.3
flask-cors==6.0.2
mysql-connector-python==9.6.0
textblob==0.19.0
PyJWT==2.11.0
python-dotenv==1.2.2
gunicorn==21.2.0